moves custom helper functions to its own dir
