+++ firmware/freedom/rules.mk
@@ -13,4 +13,4 @@ RGBLIGHT_DRIVER = ws2812
 
 SRC += matrix.c
 SRC += analog.c
-SRC += sensor_read.c
+SRC += helpers/sensor_read.c
